[nt] Move constants outside of page folder and update import paths
